+++ js/stateBox2d.js
@@ -111,8 +111,6 @@ Game.StateBox2d = Flynn.State.extend({
 
     render: function(ctx){
 
-        ctx.clearAll();
-
         var left_x = 10;
         var margin = 3;
         var i, j, x;