+++ js/stateHome.js
@@ -287,8 +287,6 @@ Game.StateHome = Flynn.State.extend({
     },
 
     render: function(ctx){
-        ctx.clearAll();
-
         var left_x = 10;
         var margin = 3;
         var i, x;